Run with 2 Arduino Devices
As the Arduino Devices increase, add the needed databases to store the data for each device. In this update, database1 refers to arduino A1 (with fetch data route: 'A1'), database2 refers to arduino A2 (with fetch data route: 'A1'). Remember to change the route in index.html file to fetch the data from the right database.